+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.4.0] - 2024-01-02
+
 ### Added
 
 - Cloud SQL policies (#12):
@@ -26,6 +28,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - `requireRootPasswordMysql`
 - SQL Server policies (#15):
   - `requireDatabaseFlagsSqlServer`
+- PostgreSQL policies (#16):
+  - `requireDatabaseFlagsPostgresql`
 
 ### Changed